Move GetWarnings and related globals to util.
This is a first step in avoiding racy accesses to strMiscWarning.

By itself this commit causes a link error because global variables relating to alerts
are only accessible in the zcashd server, but util is used in other binaries.
